Fixes to writing stats as floating point + new FVR level

+++ Assets/Scripts/Free Game/FreeGlobals.cs
@@ -214,7 +214,7 @@ public static class FreeGlobals
 
 			// Helper vars
 			TimeSpan te = gameEndTime.Subtract (gameStartTime);
-			int numMinElapsed = te.Hours * 60 + te.Minutes + (int)Math.Round ((double)te.Seconds / 60);
+			float numMinElapsed = te.Hours * 60 + te.Minutes + (int)Math.Round ((double)te.Seconds / 60);
 			if (numMinElapsed == 0)
 				numMinElapsed = 1;
